# fTsrv1/swagger.json
{
  "info": {
    "title": "fusion-swagger",
    "version": "1.0.0",
    "description": "internal-server swagger"
  },
  "openapi": "3.0.0",
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "authorization"
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/api/request/{userId}": {
      "get": {
        "summary": "GET all user's request records",
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "description": "user's DB pkey",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "tags": [
          "requests"
        ],
        "responses": {
          "200": {
            "description": "Particular user's requests"
          },
          "500": {
            "description": "error message. Probably error in parameters"
          }
        }
      }
    },
    "/api/request/": {
      "get": {
        "summary": "GET all request records",
        "parameters": [
          {
            "in": "query",
            "name": "filter",
            "allowReserved": true,
            "schema": {
              "type": "string",
              "example": {
                "type": "string",
                "status": "string"
              },
              "description": "filter for where STRINGIFIED object"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "array",
              "collectionFormat": "multi",
              "items": {
                "type": "string"
              },
              "description": "sorting parameters",
              "example": [
                "id",
                "ASC"
              ]
            }
          }
        ],
        "tags": [
          "requests"
        ],
        "responses": {
          "200": {
            "description": "Filtered and sorted list of requests"
          },
          "500": {
            "description": "error message. Probably error in parameters"
          }
        }
      },
      "put": {
        "summary": "Update request record",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/requestModel"
              }
            }
          }
        },
        "tags": [
          "requests"
        ],
        "responses": {
          "200": {
            "description": "Updated request record"
          },
          "500": {
            "description": "err message. probably invalid request body"
          }
        }
      },
      "post": {
        "summary": "Create request record",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/createRequestModel"
              }
            }
          }
        },
        "tags": [
          "requests"
        ],
        "responses": {
          "201": {
            "description": "new request created"
          },
          "500": {
            "description": "err message. probably invalid request body"
          }
        }
      }
    },
    "/api/request/{id}": {
      "delete": {
        "summary": "Delete request record",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "request db PK",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "tags": [
          "requests"
        ],
        "responses": {
          "200": {
            "description": "request deleted"
          },
          "500": {
            "description": "Error message. Probably invalid request data"
          }
        }
      }
    },
...
    "/api/tech-group/": {
      "get": {
        "summary": "GET all tech groups",
        "tags": [
          "techGroup"
        ],
        "responses": {
          "200": {
            "description": "techGroups"
          },
          "500": {
            "description": "err message"
          }
        }
      },
      "post": {
        "summary": "create new tech group",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/TechGroupModel"
              }
            },
            "description": "new tech group object"
          }
        },
        "tags": [
          "techGroup"
        ],
        "responses": {
          "200": {
            "description": "tech group created"
          },
          "500": {
            "description": "err message"
          }
        }
      },
      "put": {
        "summary": "update tech group",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "number"
            },
            "description": "db Pkey"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/TechGroupModel"
              }
            },
            "description": "new tech group object"
          }
        },
        "tags": [
          "techGroup"
        ],
        "responses": {
          "200": {
            "description": "tech group updated"
          },
          "500": {
            "description": "err message"
          }
        }
      }
    },
    "/api/tech-group/{id}": {
      "delete": {
        "summary": "delete tech group",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "number"
            },
            "description": "techGroup db pkey"
          }
        ],
        "tags": [
          "techGroup"
        ],
        "responses": {
          "200": {
            "description": "tech group deleted"
          },
          "500": {
            "description": "err message"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "requestModel": {
      "type": "object",
      "required": [
        "id"
      ],
      "properties": {
        "id": {
          "type": "number"
        },
        "title": {
          "type": "string"
        },
        "type": {
          "type": "string",
          "enum": [
            "technical",
            "vacation",
            "medical",
            "dayOff",
            "common",
            "documents"
          ]
        },
        "dateFrom": {
          "type": "string",
          "format": "date-time"
        },
        "dateTo": {
          "type": "string",
          "format": "date-time"
        },
        "dates": {
          "type": "array",
          "collectionFormat": "multi",
          "items": {
            "type": "string",
            "format": "date-time"
          }
        },
        "comment": {
          "type": "string"
        },
        "deniedComment": {
          "type": "string"
        },
        "updatedBy": {
          "type": "number"
        },
        "status": {
          "type": "string",
          "enum": [
            "wait",
            "completed",
            "denied",
            "inProgress"
          ]
        }
      }
    },
    "createRequestModel": {
      "type": "object",
      "required": [
        "title",
        "date",
        "type",
        "comment"
      ],
      "properties": {
        "title": {
          "type": "string"
        },
        "type": {
          "type": "string",
          "enum": [
            "technical",
            "vacation",
            "medical",
            "dayOff",
            "common",
            "documents"
          ]
        },
        "date": {
          "type": "string",
          "format": "date-time"
        },
        "comment": {
          "type": "string"
        },
        "from": {
          "type": "object",
          "properties": {
            "id": {
              "type": "number"
            }
          }
        }
      }
    },
...
    "TechGroupModel": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        }
      }
    },
...
  }
}